Revert "Rename file and clean up code"
This reverts commit 543c8362319983a21742c1431a70175fde7890e9.
