+++ html/render.js
@@ -4,7 +4,8 @@ import ReactDOM from 'react-dom'
 import { SceaHomePageLatestExperimentContainer } from '../src/index'
 
 const renderLatestExperimentContainer = (options, target) => {
-  ReactDOM.render(<SceaHomePageLatestExperimentContainer {...options} />, document.getElementById(target))
+    ReactDOM.render(<SceaHomePageLatestExperimentContainer {...options} />, document.getElementById(target))
 }
+
 
 export {renderLatestExperimentContainer}
